projects: adt7420-pmdz: Fix passing argument from incompatible pointer type error
projects/adt7420-pmdz/src/examples/dummy/dummy_example.c:142:67: error: passing argument 3
of 'adt7420_reg_read' from incompatible pointer type [-Wincompatible-pointer-types]
  142 |                 ret = adt7420_reg_read(adt7420, ADT7420_REG_HIST, &readval);
      |                                                                   ^~~~~~~~
      |                                                                   |
      |                                                                   uint32_t * {aka long unsigned int *}

+++ projects/adt7420-pmdz/src/examples/dummy/dummy_example.c
@@ -138,7 +138,7 @@ int example_main()
 		ret = adt7420_reg_write(adt7420, ADT7420_REG_HIST, hyst);
 		if (ret)
 			goto error_adt7420;
-		uint32_t readval = 0;
+		uint16_t readval = 0;
 		ret = adt7420_reg_read(adt7420, ADT7420_REG_HIST, &readval);
 		if (ret)
 			goto error_adt7420;
